cleaned up front-end, fixed nodemon windows functionality

+++ Applications/Week-01/01-BasicInsecureWebApp/front-end/scripts.js
@@ -18,8 +18,16 @@ function addCar(e) {
     let model = $("#model");
     let carid = $("#carid");
 
+    let makeVal = make.val();
+    let modelVal = model.val();
+
+    if(makeVal == "" || modelVal == "") {
+        alert('Make and Model cannot be blank');
+        return;
+    }
+
     if (+carid.val() === 0) {
-        jQuery.post("http://localhost:3000/api/car", { make: make.val(), model: model.val() }, function(data) {
+        jQuery.post("http://localhost:3000/api/car", { make: makeVal, model: modelVal }, function(data) {
             getCars();
         });
     } else {
@@ -34,6 +42,7 @@ function addCar(e) {
     }
 
     carid.val(0);
+    $("#car-submit").val('Add Car');
     model.val("");
     make.val("");
 }
@@ -44,11 +53,13 @@ function editCar(e) {
     let make = $("#make");
     let model = $("#model");
     let id = $("#carid");
+    
 
     let makeVal = el.data("carmake");
     let modelVal = el.data("carmodel");
     let idVal = el.data("carid");
 
+    $("#car-submit").val(`Edit Car #${idVal}`);
     make.val(makeVal);
     model.val(modelVal);
     id.val(idVal);
@@ -56,15 +67,18 @@ function editCar(e) {
 
 function delCar(e) {
     e.preventDefault();
+    
     let el = $(e.srcElement);
     let carid = el.data("carid");
-    $.ajax({
-            method: "DELETE",
-            url: "http://localhost:3000/api/car/" + carid
-        })
-        .done(function(msg) {
-            getCars();
-        });
+    if(confirm(`Are you sure you want to delete car #${carid}`)) {
+        $.ajax({
+                method: "DELETE",
+                url: "http://localhost:3000/api/car/" + carid
+            })
+            .done(function(msg) {
+                getCars();
+            });
+    }
 }
